refactor: componentizar ClassDetails.tsx em módulos separados
Divide o arquivo monolítico de 1755 linhas em 14 arquivos:
types, useClassDetails hook, ClassHero, ActivitiesTab, StudentsTab,
JamSessionsTab, ActivityFormModal (unifica create/edit), ActivityViewModal,
DeleteActivityModal, SubmissionsModal, SubmissionStatusBadge, TestCaseRow,
index barrel e ClassDetails orquestrador.
