+++ front/src/pages/classDetails/ClassDetails.tsx
@@ -1,270 +1,62 @@
-import { useEffect, useState } from "react";
 import { useParams, useNavigate, useSearchParams } from "react-router";
-import type { Class, ClassStudent } from "@/types/classes";
-import type { Student, Activity, Problem, Submission, SubmissionStatus, TestCaseResult } from "@/types";
-import ClassesService from "@/services/ClassesService";
-import { getAllStudents } from "@/services/StudentsService";
-import { getActivitiesByClass, createActivity, updateActivity, deleteActivity, getActivitySubmissions } from "@/services/ActivitiesService";
-import { getAllProblems, getProblemById } from "@/services/ProblemsServices";
-import { getResultBySubmissionId } from "@/services/SubmissionsService";
 import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
-import { DateTimePicker } from "@/components/ui/date-time-picker";
 import Loading from "@/components/Loading";
 import Notification from "@/components/Notification";
-import { ArrowLeft, UserPlus, UserMinus, Users, Search, BookOpen, Plus, X, Codesandbox, Clock, HardDrive, Calendar, MoreVertical, Pencil, Trash2, CheckCircle2, AlertCircle, XCircle, Loader2, Eye, EyeOff, Copy, Hash, Terminal, Target as TargetIcon, TestTube, FileText } from "lucide-react";
-import { useUserRole } from "@/hooks/useUserRole";
+import { ArrowLeft, BookOpen, Users, Codesandbox } from "lucide-react";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
-import { JamSessionService } from "@/services/JamSessionService";
 import JamSessionBanner from "@/components/jam/JamSessionBanner";
-import type { JamSession } from "@/types/jam";
 import ProblemViewModal from "@/components/ProblemViewModal";
-import { RichTextViewer } from "@/components/RichTextEditor";
-import { CodeViewer } from "@/components/CodeSubmission";
-import {
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableHeader,
-  TableRow,
-} from "@/components/table";
+import { useClassDetails } from "./useClassDetails";
+import ClassHero from "./ClassHero";
+import ActivitiesTab from "./ActivitiesTab";
+import StudentsTab from "./StudentsTab";
+import JamSessionsTab from "./JamSessionsTab";
+import ActivityFormModal from "./ActivityFormModal";
+import ActivityViewModal from "./ActivityViewModal";
+import DeleteActivityModal from "./DeleteActivityModal";
+import SubmissionsModal from "./SubmissionsModal";
 
 export default function ClassDetails() {
   const { id } = useParams<{ id: string }>();
   const navigate = useNavigate();
   const [searchParams] = useSearchParams();
-  const { hasAnyRole } = useUserRole();
-
   const defaultTab = searchParams.get("tab") || "activities";
 
-  const [classData, setClassData] = useState<Class | null>(null);
-  const [students, setStudents] = useState<ClassStudent[]>([]);
-  const [activities, setActivities] = useState<Activity[]>([]);
-  const [allStudents, setAllStudents] = useState<Student[]>([]);
-  const [filteredStudents, setFilteredStudents] = useState<Student[]>([]);
-  const [loading, setLoading] = useState(true);
-  const [showAddStudent, setShowAddStudent] = useState(false);
-  const [searchTerm, setSearchTerm] = useState("");
-  const [notification, setNotification] = useState<{
-    message: string;
-    type: "success" | "error";
-  } | null>(null);
-  const [showNewActivity, setShowNewActivity] = useState(false);
-  const [allProblems, setAllProblems] = useState<Problem[]>([]);
-  const [selectedProblem, setSelectedProblem] = useState<Problem | null>(null);
-  const [viewProblem, setViewProblem] = useState<Problem | null>(null);
-  const [viewActivity, setViewActivity] = useState<Activity | null>(null);
-  const [editingActivity, setEditingActivity] = useState<Activity | null>(null);
-  const [deletingActivity, setDeletingActivity] = useState<Activity | null>(null);
-  const [openMenuId, setOpenMenuId] = useState<number | null>(null);
-  const [viewSubmissionsActivity, setViewSubmissionsActivity] = useState<Activity | null>(null);
-  const [jamSessions, setJamSessions] = useState<JamSession[]>([]);
-  const [activeJam, setActiveJam] = useState<JamSession | null>(null);
-
-  useEffect(() => {
-    if (id) {
-      loadClassData();
-      loadClassStudents();
-      loadAllStudents();
-      loadClassActivities();
-      loadProblems();
-      loadJamSessions();
-    }
-  }, [id]);
-
-  useEffect(() => {
-    filterAvailableStudents();
-  }, [searchTerm, allStudents, students]);
-
-  useEffect(() => {
-    const handleClickOutside = (event: MouseEvent) => {
-      const target = event.target as HTMLElement;
-      if (!target.closest('.activity-menu')) {
-        setOpenMenuId(null);
-      }
-    };
-
-    if (openMenuId !== null) {
-      setTimeout(() => {
-        document.addEventListener("mousedown", handleClickOutside);
-      }, 0);
-    }
-
-    return () => {
-      document.removeEventListener("mousedown", handleClickOutside);
-    };
-  }, [openMenuId]);
-
-  const loadClassData = async () => {
-    try {
-      const data = await ClassesService.getClassById(Number(id));
-      setClassData(data);
-    } catch (error) {
-      console.error("Erro ao carregar turma:", error);
-      showNotification("Erro ao carregar turma", "error");
-    }
-  };
-
-  const loadClassStudents = async () => {
-    try {
-      const data = await ClassesService.getClassStudents(Number(id));
-      setStudents(data);
-    } catch (error) {
-      console.error("Erro ao carregar alunos:", error);
-      showNotification("Erro ao carregar alunos", "error");
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  const loadAllStudents = async () => {
-    try {
-      const data = await getAllStudents();
-      console.log("Alunos carregados do banco:", data);
-      setAllStudents(data);
-    } catch (error) {
-      console.error("Erro ao carregar lista de alunos:", error);
-    }
-  };
-
-  const loadClassActivities = async () => {
-    try {
-      const data = await getActivitiesByClass(Number(id));
-      setActivities(data);
-    } catch (error) {
-      console.error("Erro ao carregar atividades:", error);
-    }
-  };
-
-  const loadProblems = async () => {
-    try {
-      const data = await getAllProblems();
-      setAllProblems(data);
-    } catch (error) {
-      console.error("Erro ao carregar problemas:", error);
-    }
-  };
-
-  const loadJamSessions = async () => {
-    try {
-      const sessions = await JamSessionService.getByTurma(Number(id));
-      setJamSessions(sessions);
-      const active = await JamSessionService.getActiveForTurma(Number(id));
-      setActiveJam(active);
-    } catch (error) {
-      console.error("Erro ao carregar jam sessions:", error);
-    }
-  };
-
-  const filterAvailableStudents = () => {
-    const enrolledIds = students.map((s) => s.studentId);
-    let available = allStudents.filter((s) => !enrolledIds.includes(s.id));
-
-    if (searchTerm.trim()) {
-      available = available.filter(
-        (s) =>
-          s.name.toLowerCase().includes(searchTerm.toLowerCase()) ||
-          s.email.toLowerCase().includes(searchTerm.toLowerCase())
-      );
-    }
-
-    setFilteredStudents(available);
-  };
-
-  const handleAddStudent = async (studentId: number) => {
-    try {
-      await ClassesService.addStudentToClass(Number(id), { studentId });
-      showNotification("Aluno adicionado com sucesso!", "success");
-      loadClassStudents();
-      setSearchTerm("");
-    } catch (error) {
-      console.error("Erro ao adicionar aluno:", error);
-      showNotification("Erro ao adicionar aluno", "error");
-    }
-  };
-
-  const handleRemoveStudent = async (studentId: number) => {
-    if (!confirm("Tem certeza que deseja remover este aluno da turma?")) return;
-
-    try {
-      await ClassesService.removeStudentFromClass(Number(id), studentId);
-      showNotification("Aluno removido com sucesso!", "success");
-      loadClassStudents();
-    } catch (error) {
-      console.error("Erro ao remover aluno:", error);
-      showNotification("Erro ao remover aluno", "error");
-    }
-  };
-
-  const showNotification = (message: string, type: "success" | "error") => {
-    setNotification({ message, type });
-    setTimeout(() => setNotification(null), 3000);
-  };
-
-  const handleCreateActivity = async (formData: { problema_id: number; data_entrega: string }) => {
-    try {
-      const result = await createActivity({
-        ...formData,
-        turma_id: Number(id),
-      });
-
-      if (result) {
-        showNotification("Atividade criada com sucesso!", "success");
-        setShowNewActivity(false);
-        setSelectedProblem(null);
-        loadClassActivities();
-      } else {
-        showNotification("Erro ao criar atividade", "error");
-      }
-    } catch (error) {
-      console.error("Erro ao criar atividade:", error);
-      showNotification("Erro ao criar atividade", "error");
-    }
-  };
-
-  const handleUpdateActivity = async (formData: { problema_id: number; data_entrega: string }) => {
-    if (!editingActivity) return;
-
-    try {
-      const result = await updateActivity(editingActivity.id, {
-        ...formData,
-        turma_id: Number(id),
-      });
-
-      if (result) {
-        showNotification("Atividade atualizada com sucesso!", "success");
-        setEditingActivity(null);
-        setSelectedProblem(null);
-        loadClassActivities();
-      } else {
-        showNotification("Erro ao atualizar atividade", "error");
-      }
-    } catch (error) {
-      console.error("Erro ao atualizar atividade:", error);
-      showNotification("Erro ao atualizar atividade", "error");
-    }
-  };
-
-  const handleDeleteActivity = async () => {
-    if (!deletingActivity) return;
-
-    try {
-      const success = await deleteActivity(deletingActivity.id);
-
-      if (success) {
-        showNotification("Atividade deletada com sucesso!", "success");
-        setDeletingActivity(null);
-        loadClassActivities();
-      } else {
-        showNotification("Erro ao deletar atividade", "error");
-      }
-    } catch (error) {
-      console.error("Erro ao deletar atividade:", error);
-      showNotification("Erro ao deletar atividade", "error");
-    }
-  };
+  const {
+    classData,
+    students,
+    activities,
+    allProblems,
+    filteredStudents,
+    jamSessions,
+    activeJam,
+    loading,
+    notification,
+    showAddStudent,
+    setShowAddStudent,
+    searchTerm,
+    setSearchTerm,
+    showNewActivity,
+    setShowNewActivity,
+    viewProblem,
+    setViewProblem,
+    viewActivity,
+    setViewActivity,
+    editingActivity,
+    setEditingActivity,
+    deletingActivity,
+    setDeletingActivity,
+    openMenuId,
+    setOpenMenuId,
+    viewSubmissionsActivity,
+    setViewSubmissionsActivity,
+    setNotification,
+    handleAddStudent,
+    handleRemoveStudent,
+    handleCreateActivity,
+    handleUpdateActivity,
+    handleDeleteActivity,
+  } = useClassDetails(id);
 
   if (loading) return <Loading />;
 
@@ -291,7 +83,6 @@ export default function ClassDetails() {
         />
       )}
 
-      {/* Back link */}
       <button
         onClick={() => navigate("/classes")}
         className="inline-flex items-center gap-1 text-sm font-medium transition-colors"
@@ -301,41 +92,14 @@ export default function ClassDetails() {
         Voltar para Turmas
       </button>
 
-      {/* Hero Header */}
-      <div
-        className="relative rounded-2xl px-8 py-10 overflow-hidden"
-        style={{ background: "linear-gradient(135deg, #0d9488 0%, #065f46 100%)" }}
-      >
-        <div className="absolute -top-6 -right-6 h-32 w-32 rounded-full bg-white opacity-10" />
-        <div className="absolute bottom-4 left-1/3 h-20 w-20 rounded-full bg-white opacity-[0.07]" />
+      <ClassHero
+        classData={classData}
+        activityCount={activities.length}
+        studentCount={students.length}
+      />
 
-        <div className="relative flex flex-col sm:flex-row sm:items-end sm:justify-between gap-4">
-          <div>
-            <h1 className="text-3xl font-extrabold tracking-tight text-white flex items-center gap-3">
-              <BookOpen className="w-8 h-8" />
-              {classData.nome}
-            </h1>
-            {classData.teacherName && (
-              <p className="text-teal-100 text-sm mt-2">
-                Professor: {classData.teacherName}
-              </p>
-            )}
-            <div className="flex items-center gap-4 mt-3">
-              <span className="text-sm font-medium px-3 py-1 rounded-full bg-white/15 text-white">
-                {activities.length} atividade{activities.length !== 1 ? "s" : ""}
-              </span>
-              <span className="text-sm font-medium px-3 py-1 rounded-full bg-white/15 text-white">
-                {students.length} aluno{students.length !== 1 ? "s" : ""}
-              </span>
-            </div>
-          </div>
-        </div>
-      </div>
-
-      {/* Banner de Jam Session ativa */}
       {activeJam && <JamSessionBanner session={activeJam} />}
 
-      {/* Tabs: Atividades, Alunos e Jam Sessions */}
       <Tabs defaultValue={defaultTab} className="w-full">
         <TabsList className="grid w-full grid-cols-3 mb-6">
           <TabsTrigger value="activities" className="flex items-center gap-2">
@@ -352,347 +116,60 @@ export default function ClassDetails() {
           </TabsTrigger>
         </TabsList>
 
-        {/* Aba de Atividades */}
         <TabsContent value="activities">
-          <div className="bg-white p-6 rounded-2xl border border-stone-200 shadow-sm">
-            <div className="flex justify-between items-center mb-4">
-              <h2 className="text-lg font-semibold text-stone-800">Atividades da Turma</h2>
-              {hasAnyRole(["professor", "admin"]) && (
-                <Button onClick={() => setShowNewActivity(true)}
-                  className="text-white rounded-xl hover:opacity-90 font-semibold transition-opacity h-11 px-6" style={{ backgroundColor: "#0d9488" }}>
-                  <Plus className="w-4 h-4 mr-2" />
-                  Nova Atividade
-                </Button>
-              )}
-            </div>
-
-            <div className="space-y-3">
-              {activities.length === 0 ? (
-                <div className="flex flex-col items-center justify-center py-16">
-                  <div className="h-16 w-16 rounded-2xl flex items-center justify-center mb-4" style={{ backgroundColor: "#ccfbf1" }}>
-                    <BookOpen className="w-7 h-7" style={{ color: "#0d9488" }} />
-                  </div>
-                  <p className="text-sm font-medium text-stone-600">Nenhuma atividade cadastrada</p>
-                  <p className="text-xs text-stone-400 mt-1">Crie a primeira atividade para esta turma</p>
-                </div>
-              ) : (
-                activities.map((activity) => {
-                  const dueDate = new Date(activity.dueDate);
-                  const now = new Date();
-                  const isOverdue = dueDate < now;
-                  const isPending = !isOverdue;
-                  const problem = allProblems.find(p => p.id === activity.problemId);
-
-                  return (
-                    <div
-                      key={activity.id}
-                      className={`p-4 pb-14 rounded-xl border-l-4 transition-all hover:shadow-md relative ${isOverdue
-                        ? "bg-red-50 border-red-500"
-                        : isPending
-                          ? "bg-amber-50 border-amber-500"
-                          : "bg-emerald-50 border-emerald-500"
-                        }`}
-                    >
-                      <div className="flex justify-between items-start">
-                        <div className="flex-1">
-                          <h3 className="font-semibold text-lg">
-                            {problem?.title || "Atividade"}
-                          </h3>
-                          <p className="text-sm text-stone-600 mt-1">
-                            Prazo: {dueDate.toLocaleDateString("pt-BR")} às{" "}
-                            {dueDate.toLocaleTimeString("pt-BR", {
-                              hour: "2-digit",
-                              minute: "2-digit",
-                            })}
-                          </p>
-                        </div>
-                        {hasAnyRole(["professor", "admin"]) && (
-                          <div className="relative ml-4 activity-menu">
-                            <button
-                              onClick={(e) => {
-                                e.stopPropagation();
-                                setOpenMenuId(openMenuId === activity.id ? null : activity.id);
-                              }}
-                              className="p-2 rounded-full hover:bg-stone-100"
-                            >
-                              <MoreVertical size={20} />
-                            </button>
-
-                            {openMenuId === activity.id && (
-                              <div
-                                className="absolute right-0 mt-2 w-48 bg-white rounded-md shadow-lg z-50 border border-stone-200"
-                                onClick={(e) => e.stopPropagation()}
-                              >
-                                <div className="py-1">
-                                  <button
-                                    onClick={(e) => {
-                                      e.stopPropagation();
-                                      setEditingActivity(activity);
-                                      setOpenMenuId(null);
-                                    }}
-                                    className="flex items-center w-full text-left px-4 py-2 text-sm text-stone-700 hover:bg-stone-100"
-                                  >
-                                    <Pencil className="mr-2 h-4 w-4" />
-                                    <span>Editar</span>
-                                  </button>
-                                  <button
-                                    onClick={(e) => {
-                                      e.stopPropagation();
-                                      setDeletingActivity(activity);
-                                      setOpenMenuId(null);
-                                    }}
-                                    className="flex items-center w-full text-left px-4 py-2 text-sm text-red-600 hover:bg-stone-100"
-                                  >
-                                    <Trash2 className="mr-2 h-4 w-4" />
-                                    <span>Apagar</span>
-                                  </button>
-                                </div>
-                              </div>
-                            )}
-                          </div>
-                        )}
-                      </div>
-
-                      {/* Botões de ação */}
-                      <div className="absolute bottom-3 left-3 flex gap-2">
-                        <Button
-                          type="button"
-                          variant="outline"
-                          size="sm"
-                          onClick={(e) => {
-                            e.stopPropagation();
-                            setViewActivity(activity);
-                          }}
-                          className="h-8 text-xs gap-2"
-                        >
-                          <Eye className="w-3.5 h-3.5" />
-                          Ver Problema
-                        </Button>
-                        {hasAnyRole(["professor", "admin"]) && (
-                          <Button
-                            type="button"
-                            variant="outline"
-                            size="sm"
-                            onClick={(e) => {
-                              e.stopPropagation();
-                              setViewSubmissionsActivity(activity);
-                            }}
-                            className="h-8 text-xs gap-2"
-                          >
-                            <FileText className="w-3.5 h-3.5" />
-                            Ver Submissões
-                          </Button>
-                        )}
-                      </div>
-
-                      <span
-                        className={`absolute bottom-3 right-10 px-2.5 py-1 rounded-lg text-xs font-medium ${isOverdue
-                          ? "bg-red-100 text-red-700"
-                          : isPending
-                            ? "bg-amber-100 text-amber-700"
-                            : "bg-emerald-100 text-emerald-700"
-                          }`}
-                      >
-                        {isOverdue
-                          ? "Atrasada"
-                          : isPending
-                            ? "Pendente"
-                            : "Concluída"}
-                      </span>
-                    </div>
-                  );
-                })
-              )}
-            </div>
-          </div>
+          <ActivitiesTab
+            activities={activities}
+            allProblems={allProblems}
+            openMenuId={openMenuId}
+            onNewActivity={() => setShowNewActivity(true)}
+            onViewActivity={(a) => setViewActivity(a)}
+            onEditActivity={(a) => { setEditingActivity(a); setOpenMenuId(null); }}
+            onDeleteActivity={(a) => { setDeletingActivity(a); setOpenMenuId(null); }}
+            onViewSubmissions={(a) => setViewSubmissionsActivity(a)}
+            onToggleMenu={(activityId) => setOpenMenuId(openMenuId === activityId ? null : activityId)}
+          />
         </TabsContent>
 
-        {/* Aba de Alunos */}
         <TabsContent value="students">
-          <div className="bg-white p-6 rounded-2xl border border-stone-200 shadow-sm">
-            <div className="flex justify-between items-center mb-4">
-              <h2 className="text-lg font-semibold text-stone-800">Alunos Matriculados</h2>
-              {hasAnyRole(["professor", "admin"]) && (
-                <Button onClick={() => setShowAddStudent(!showAddStudent)}>
-                  <UserPlus className="w-4 h-4 mr-2" />
-                  Adicionar Aluno
-                </Button>
-              )}
-            </div>
-
-            {/* Formulário de adicionar aluno */}
-            {showAddStudent && (
-              <div className="mb-6 p-4 bg-stone-50 rounded-lg">
-                <div className="mb-4">
-                  <div className="relative">
-                    <Search className="absolute left-3 top-3 w-4 h-4 text-stone-400" />
-                    <Input
-                      type="text"
-                      placeholder="Buscar alunos disponíveis..."
-                      value={searchTerm}
-                      onChange={(e) => setSearchTerm(e.target.value)}
-                      className="pl-10"
-                    />
-                  </div>
-                </div>
-
-                <div className="max-h-64 overflow-y-auto space-y-2">
-                  {filteredStudents.length === 0 ? (
-                    <p className="text-stone-500 text-center py-4">
-                      {searchTerm
-                        ? "Nenhum aluno encontrado"
-                        : "Todos os alunos já estão matriculados"}
-                    </p>
-                  ) : (
-                    filteredStudents.map((student) => (
-                      <div
-                        key={student.id}
-                        className="flex justify-between items-center p-3 bg-white rounded border hover:border-teal-300 transition-colors"
-                      >
-                        <div>
-                          <p className="font-medium">{student.name}</p>
-                          <p className="text-sm text-stone-600">
-                            {student.email}
-                          </p>
-                        </div>
-                        <Button
-                          size="sm"
-                          onClick={() => handleAddStudent(student.id)}
-                        >
-                          <UserPlus className="w-4 h-4" />
-                        </Button>
-                      </div>
-                    ))
-                  )}
-                </div>
-              </div>
-            )}
-
-            {/* Lista de alunos matriculados */}
-            <div className="space-y-2">
-              {students.length === 0 ? (
-                <div className="flex flex-col items-center justify-center py-16">
-                  <div className="h-16 w-16 rounded-2xl flex items-center justify-center mb-4" style={{ backgroundColor: "#ccfbf1" }}>
-                    <Users className="w-7 h-7" style={{ color: "#0d9488" }} />
-                  </div>
-                  <p className="text-sm font-medium text-stone-600">Nenhum aluno matriculado</p>
-                  <p className="text-xs text-stone-400 mt-1">Adicione alunos a esta turma</p>
-                </div>
-              ) : (
-                students.map((student) => (
-                  <div
-                    key={student.id}
-                    className="flex justify-between items-center p-4 bg-stone-50 rounded-lg hover:bg-stone-100 transition-colors"
-                  >
-                    <div>
-                      <p className="font-medium">{student.studentName}</p>
-                      <p className="text-sm text-stone-600">
-                        {student.studentEmail}
-                      </p>
-                      <p className="text-xs text-stone-500 mt-1">
-                        Matriculado em:{" "}
-                        {new Date(student.enrolledAt).toLocaleDateString(
-                          "pt-BR"
-                        )}
-                      </p>
-                    </div>
-                    {hasAnyRole(["professor", "admin"]) && (
-                      <Button
-                        size="sm"
-                        variant="outline"
-                        onClick={() => handleRemoveStudent(student.studentId)}
-                        className="text-red-600 hover:text-red-700 hover:bg-red-50"
-                      >
-                        <UserMinus className="w-4 h-4" />
-                      </Button>
-                    )}
-                  </div>
-                ))
-              )}
-            </div>
-          </div>
+          <StudentsTab
+            students={students}
+            showAddStudent={showAddStudent}
+            searchTerm={searchTerm}
+            filteredStudents={filteredStudents}
+            onToggleAddStudent={() => setShowAddStudent(!showAddStudent)}
+            onSearchTermChange={setSearchTerm}
+            onAddStudent={handleAddStudent}
+            onRemoveStudent={handleRemoveStudent}
+          />
         </TabsContent>
 
-        {/* Aba de Jam Sessions */}
         <TabsContent value="jam">
-          <div className="bg-white p-6 rounded-2xl border border-stone-200 shadow-sm">
-            <div className="flex justify-between items-center mb-4">
-              <h2 className="text-lg font-semibold text-stone-800">Jam Sessions</h2>
-              {hasAnyRole(["professor", "admin"]) && (
-                <Button
-                  onClick={() => navigate(`/jam/create/${id}`)}
-                  className="flex items-center gap-2"
-                >
-                  <Plus className="w-4 h-4" />
-                  Nova Sessão
-                </Button>
-              )}
-            </div>
-
-            {jamSessions.length === 0 ? (
-              <div className="flex flex-col items-center justify-center py-16">
-                <div className="h-16 w-16 rounded-2xl flex items-center justify-center mb-4" style={{ backgroundColor: "#ccfbf1" }}>
-                  <Codesandbox className="w-7 h-7" style={{ color: "#0d9488" }} />
-                </div>
-                <p className="text-sm font-medium text-stone-600">Nenhuma jam session criada</p>
-                <p className="text-xs text-stone-400 mt-1">Crie uma sessão para começar</p>
-              </div>
-            ) : (
-              <div className="space-y-3">
-                {jamSessions.map((jam) => (
-                  <div
-                    key={jam.id}
-                    onClick={() => navigate(`/jam/${jam.id}`)}
-                    className="flex items-center justify-between rounded-lg border border-stone-200 p-4 cursor-pointer hover:bg-stone-50 transition-colors"
-                  >
-                    <div>
-                      <h3 className="font-semibold text-stone-800">{jam.titulo}</h3>
-                      <p className="text-sm text-stone-500">
-                        Problema: {jam.problema?.titulo || "—"}
-                        {jam.tempo_limite && ` · ${jam.tempo_limite} min`}
-                      </p>
-                    </div>
-                    <div className="flex items-center gap-3">
-                      <span className="text-xs text-stone-400">
-                        {jam.participants?.length || 0} participantes
-                      </span>
-                      <span
-                        className={`rounded-full px-3 py-1 text-xs font-medium ${
-                          jam.status === "active"
-                            ? "bg-green-100 text-green-700"
-                            : jam.status === "waiting"
-                            ? "bg-yellow-100 text-yellow-700"
-                            : "bg-stone-100 text-stone-600"
-                        }`}
-                      >
-                        {jam.status === "active"
-                          ? "Ativa"
-                          : jam.status === "waiting"
-                          ? "Aguardando"
-                          : "Encerrada"}
-                      </span>
-                    </div>
-                  </div>
-                ))}
-              </div>
-            )}
-          </div>
+          <JamSessionsTab jamSessions={jamSessions} classId={id!} />
         </TabsContent>
       </Tabs>
 
       {showNewActivity && (
-        <NewActivityModal
+        <ActivityFormModal
+          key="new"
           isOpen={showNewActivity}
-          onClose={() => {
-            setShowNewActivity(false);
-            setSelectedProblem(null);
-          }}
+          onClose={() => setShowNewActivity(false)}
           onSave={handleCreateActivity}
           problems={allProblems}
           onViewProblem={(problem) => setViewProblem(problem)}
         />
       )}
 
+      {editingActivity && (
+        <ActivityFormModal
+          key={editingActivity.id}
+          isOpen={!!editingActivity}
+          onClose={() => setEditingActivity(null)}
+          onSave={handleUpdateActivity}
+          activity={editingActivity}
+          problems={allProblems}
+          onViewProblem={(problem) => setViewProblem(problem)}
+        />
+      )}
 
       {viewActivity && (
         <ActivityViewModal
@@ -700,20 +177,6 @@ export default function ClassDetails() {
           activity={viewActivity}
           problem={allProblems.find(p => p.id === viewActivity.problemId) || null}
           onClose={() => setViewActivity(null)}
-        />
-      )}
-
-      {editingActivity && (
-        <EditActivityModal
-          isOpen={!!editingActivity}
-          onClose={() => {
-            setEditingActivity(null);
-            setSelectedProblem(null);
-          }}
-          onSave={handleUpdateActivity}
-          activity={editingActivity}
-          problems={allProblems}
-          onViewProblem={(problem) => setViewProblem(problem)}
         />
       )}
 
@@ -738,1018 +201,6 @@ export default function ClassDetails() {
         activity={viewSubmissionsActivity}
         classId={Number(id)}
       />
-
     </div>
-  );
-}
-
-interface NewActivityModalProps {
-  isOpen: boolean;
-  onClose: () => void;
-  onSave: (data: { problema_id: number; data_entrega: string }) => void;
-  problems: Problem[];
-  onViewProblem: (problem: Problem) => void;
-}
-
-function NewActivityModal({ isOpen, onClose, onSave, problems, onViewProblem }: NewActivityModalProps) {
-  const [selectedProblemId, setSelectedProblemId] = useState<number | null>(null);
-  const [dueDate, setDueDate] = useState<Date | null>(null);
-  const [searchTerm, setSearchTerm] = useState("");
-
-  const filteredProblems = problems.filter(problem =>
-    problem.title.toLowerCase().includes(searchTerm.toLowerCase())
-  );
-
-  const selectedProblem = problems.find(p => p.id === selectedProblemId);
-
-  const handleSubmit = (e: React.FormEvent) => {
-    e.preventDefault();
-    if (!selectedProblemId || !dueDate) {
-      return;
-    }
-    const year = dueDate.getFullYear();
-    const month = String(dueDate.getMonth() + 1).padStart(2, '0');
-    const day = String(dueDate.getDate()).padStart(2, '0');
-    const hours = String(dueDate.getHours()).padStart(2, '0');
-    const minutes = String(dueDate.getMinutes()).padStart(2, '0');
-    const seconds = String(dueDate.getSeconds()).padStart(2, '0');
-    const formattedDate = `${year}-${month}-${day} ${hours}:${minutes}:${seconds}`;
-    onSave({
-      problema_id: selectedProblemId,
-      data_entrega: formattedDate,
-    });
-  };
-
-  if (!isOpen) return null;
-
-  return (
-    <div className="fixed inset-0 bg-black/30 backdrop-blur-sm flex items-center justify-center z-50 p-4">
-      <div className="bg-white rounded-2xl shadow-xl max-w-3xl w-full max-h-[90vh] overflow-y-auto">
-        <div className="p-6">
-          <div className="flex items-center justify-between border-b border-stone-200 pb-4">
-            <h2 className="text-xl font-bold text-stone-900">
-              {selectedProblem ? selectedProblem.title : "Nova Atividade"}
-            </h2>
-            <button
-              onClick={onClose}
-              className="text-stone-400 hover:text-stone-600 transition-colors"
-            >
-              <X className="w-5 h-5" />
-            </button>
-          </div>
-
-          <form onSubmit={handleSubmit} className="space-y-6 mt-6">
-            <div>
-              <Label htmlFor="search-problem">Selecione um Problema *</Label>
-              <div className="relative mt-2">
-                <Search className="absolute left-3 top-3 w-4 h-4 text-stone-400" />
-                <Input
-                  id="search-problem"
-                  placeholder="Buscar problemas..."
-                  value={searchTerm}
-                  onChange={(e) => setSearchTerm(e.target.value)}
-                  className="pl-10"
-                />
-              </div>
-
-              <div className="mt-3 max-h-64 overflow-y-auto border border-stone-200 rounded-xl">
-                {filteredProblems.length === 0 ? (
-                  <p className="text-center text-stone-500 py-8">Nenhum problema encontrado</p>
-                ) : (
-                  filteredProblems.map((problem) => (
-                    <div
-                      key={problem.id}
-                      className={`p-3 border-b last:border-b-0 cursor-pointer hover:bg-stone-50 transition-colors ${selectedProblemId === problem.id ? "bg-teal-50 border-l-4 border-l-teal-600" : ""
-                        }`}
-                      onClick={() => setSelectedProblemId(problem.id)}
-                    >
-                      <div className="flex justify-between items-start">
-                        <div className="flex items-center flex-1">
-                          <div className="flex-shrink-0 mr-3">
-                            <Codesandbox size={24} className="text-stone-500" />
-                          </div>
-                          <div className="flex-1">
-                            <h4 className="font-medium text-stone-900">{problem.title}</h4>
-                            <div className="flex gap-3 mt-1 text-xs text-stone-600">
-                              <span>Tempo: {problem.timeLimitMs}ms</span>
-                              <span>Memória: {problem.memoryLimitKb}KB</span>
-                            </div>
-                          </div>
-                        </div>
-                        <Button
-                          type="button"
-                          variant="outline"
-                          size="sm"
-                          onClick={(e) => {
-                            e.stopPropagation();
-                            onViewProblem(problem);
-                          }}
-                          className="ml-2"
-                        >
-                          Visualizar
-                        </Button>
-                      </div>
-                    </div>
-                  ))
-                )}
-              </div>
-
-              {selectedProblem && (
-                <div className="mt-3 p-3 bg-teal-50 border border-teal-200 rounded-lg">
-                  <p className="text-sm text-teal-900">
-                    <strong>Problema selecionado:</strong> {selectedProblem.title}
-                  </p>
-                </div>
-              )}
-            </div>
-
-            <div>
-              <Label htmlFor="due-date">Data de Entrega *</Label>
-              <DateTimePicker
-                selected={dueDate}
-                onChange={setDueDate}
-                placeholderText="dd/mm/aaaa às hh:mm"
-                required
-                className="mt-2"
-              />
-            </div>
-
-            <div className="flex gap-3 pt-4">
-              <button
-                type="button"
-                onClick={onClose}
-                className="flex-1 px-4 py-2 border border-stone-300 rounded-lg text-stone-700 hover:bg-stone-50 font-medium transition-colors"
-              >
-                Cancelar
-              </button>
-              <button
-                type="submit"
-                disabled={!selectedProblemId || !dueDate}
-                className="flex-1 px-4 py-2 text-white rounded-lg hover:opacity-90 font-medium transition-opacity disabled:opacity-50 disabled:cursor-not-allowed" style={{ background: "linear-gradient(135deg, #0d9488 0%, #065f46 100%)" }}
-              >
-                Criar Atividade
-              </button>
-            </div>
-          </form>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-interface ActivityViewModalProps {
-  isOpen: boolean;
-  onClose: () => void;
-  activity: Activity;
-  problem: Problem | null;
-}
-
-function ActivityViewModal({ isOpen, onClose, activity, problem }: ActivityViewModalProps) {
-  const { hasAnyRole } = useUserRole();
-  if (!isOpen || !problem) return null;
-
-  const dueDate = new Date(activity.dueDate);
-
-  return (
-    <div className="fixed inset-0 bg-black/30 backdrop-blur-sm flex items-center justify-center z-50 p-4">
-      <div className="bg-white rounded-2xl shadow-xl max-w-4xl w-full max-h-[90vh] overflow-y-auto">
-        <div className="p-6">
-          <div className="flex items-center justify-between border-b border-stone-200 pb-4">
-            <h2 className="text-xl font-bold text-stone-900 flex items-center">
-              <Codesandbox className="w-6 h-6 mr-3 text-stone-600" />
-              {problem.title}
-            </h2>
-            <button
-              onClick={onClose}
-              className="text-stone-400 hover:text-stone-600 transition-colors"
-            >
-              <X className="w-5 h-5" />
-            </button>
-          </div>
-
-          <div className="space-y-6">
-            <div className="bg-teal-50 border border-teal-200 rounded-lg p-4">
-              <h3 className="font-semibold mb-2 flex items-center text-teal-900">
-                <Calendar className="w-5 h-5 mr-2" />
-                Data de Entrega
-              </h3>
-              <p className="text-lg font-medium text-teal-900">
-                {dueDate.toLocaleDateString("pt-BR", {
-                  day: "2-digit",
-                  month: "long",
-                  year: "numeric",
-                })} às {dueDate.toLocaleTimeString("pt-BR", {
-                  hour: "2-digit",
-                  minute: "2-digit",
-                })}
-              </p>
-            </div>
-
-            <div>
-              <h3 className="font-semibold mb-2">Enunciado</h3>
-              <div className="bg-stone-50 p-4 rounded-lg">
-                <RichTextViewer value={problem.statement} />
-              </div>
-            </div>
-
-            <div className="grid grid-cols-2 gap-4">
-              <div>
-                <h3 className="font-semibold mb-2">Tempo Limite</h3>
-                <div className="flex items-center gap-2">
-                  <Clock className="w-4 h-4" />
-                  <span>{problem.timeLimitMs} milissegundos</span>
-                </div>
-              </div>
-              <div>
-                <h3 className="font-semibold mb-2">Memória Limite</h3>
-                <div className="flex items-center gap-2">
-                  <HardDrive className="w-4 h-4" />
-                  <span>{problem.memoryLimitKb} KB</span>
-                </div>
-              </div>
-            </div>
-
-            {problem.testCases && problem.testCases.length > 0 && (
-              <div>
-                <h3 className="font-semibold mb-4">Casos de Teste</h3>
-                <div className="space-y-4">
-                  {problem.testCases
-                    .filter(testCase => !testCase.private || hasAnyRole(['professor', 'admin']))
-                    .map((testCase, index) => (
-                      <div key={index} className="border rounded-lg p-4">
-                        <div className="flex justify-between items-center mb-3">
-                          <h4 className="font-medium">Caso {index + 1}</h4>
-                          {testCase.private && (
-                            <span className="px-2 py-1 bg-stone-200 text-stone-700 text-xs rounded">
-                              Privado
-                            </span>
-                          )}
-                        </div>
-                        <div className="space-y-3">
-                          <div>
-                            <Label className="text-xs font-medium">Entrada</Label>
-                            <div className="bg-stone-50 p-3 rounded mt-1">
-                              <pre className="text-sm">{testCase.input}</pre>
-                            </div>
-                          </div>
-                          <div>
-                            <Label className="text-xs font-medium">Saída Esperada</Label>
-                            <div className="bg-stone-50 p-3 rounded mt-1">
-                              <pre className="text-sm">{testCase.expectedOutput}</pre>
-                            </div>
-                          </div>
-                        </div>
-                      </div>
-                    ))}
-                </div>
-              </div>
-            )}
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-interface EditActivityModalProps {
-  isOpen: boolean;
-  onClose: () => void;
-  onSave: (data: { problema_id: number; data_entrega: string }) => void;
-  activity: Activity;
-  problems: Problem[];
-  onViewProblem: (problem: Problem) => void;
-}
-
-function EditActivityModal({ isOpen, onClose, onSave, activity, problems, onViewProblem }: EditActivityModalProps) {
-  const [selectedProblemId, setSelectedProblemId] = useState<number | null>(activity.problemId);
-  const [dueDate, setDueDate] = useState<Date | null>(new Date(activity.dueDate.replace(' ', 'T')));
-  const [searchTerm, setSearchTerm] = useState("");
-
-  useEffect(() => {
-    setSelectedProblemId(activity.problemId);
-    const parsedDate = new Date(activity.dueDate.replace(' ', 'T'));
-    setDueDate(parsedDate);
-    setSearchTerm("");
-  }, [activity, isOpen]);
-
-  const filteredProblems = problems.filter(problem =>
-    problem.title.toLowerCase().includes(searchTerm.toLowerCase())
-  );
-
-  const selectedProblem = problems.find(p => p.id === selectedProblemId);
-
-  const handleSubmit = (e: React.FormEvent) => {
-    e.preventDefault();
-    if (!selectedProblemId || !dueDate) {
-      return;
-    }
-    const year = dueDate.getFullYear();
-    const month = String(dueDate.getMonth() + 1).padStart(2, '0');
-    const day = String(dueDate.getDate()).padStart(2, '0');
-    const hours = String(dueDate.getHours()).padStart(2, '0');
-    const minutes = String(dueDate.getMinutes()).padStart(2, '0');
-    const seconds = String(dueDate.getSeconds()).padStart(2, '0');
-    const formattedDate = `${year}-${month}-${day} ${hours}:${minutes}:${seconds}`;
-    onSave({
-      problema_id: selectedProblemId,
-      data_entrega: formattedDate,
-    });
-  };
-
-  if (!isOpen) return null;
-
-  return (
-    <div className="fixed inset-0 bg-black/30 backdrop-blur-sm flex items-center justify-center z-50 p-4">
-      <div className="bg-white rounded-2xl shadow-xl max-w-3xl w-full max-h-[90vh] overflow-y-auto">
-        <div className="p-6">
-          <div className="flex items-center justify-between border-b border-stone-200 pb-4">
-            <h2 className="text-xl font-bold text-stone-900">Editar Atividade</h2>
-            <button
-              onClick={onClose}
-              className="text-stone-400 hover:text-stone-600 transition-colors"
-            >
-              <X className="w-5 h-5" />
-            </button>
-          </div>
-
-          <form onSubmit={handleSubmit} className="space-y-6 mt-6">
-            <div>
-              <Label htmlFor="search-problem-edit">Selecione um Problema *</Label>
-              <div className="relative mt-2">
-                <Search className="absolute left-3 top-3 w-4 h-4 text-stone-400" />
-                <Input
-                  id="search-problem-edit"
-                  placeholder="Buscar problemas..."
-                  value={searchTerm}
-                  onChange={(e) => setSearchTerm(e.target.value)}
-                  className="pl-10"
-                />
-              </div>
-
-              <div className="mt-3 max-h-64 overflow-y-auto border border-stone-200 rounded-xl">
-                {filteredProblems.length === 0 ? (
-                  <p className="text-center text-stone-500 py-8">Nenhum problema encontrado</p>
-                ) : (
-                  filteredProblems.map((problem) => (
-                    <div
-                      key={problem.id}
-                      className={`p-3 border-b last:border-b-0 cursor-pointer hover:bg-stone-50 transition-colors ${selectedProblemId === problem.id ? "bg-teal-50 border-l-4 border-l-teal-600" : ""
-                        }`}
-                      onClick={() => setSelectedProblemId(problem.id)}
-                    >
-                      <div className="flex justify-between items-start">
-                        <div className="flex items-center flex-1">
-                          <div className="flex-shrink-0 mr-3">
-                            <Codesandbox size={24} className="text-stone-500" />
-                          </div>
-                          <div className="flex-1">
-                            <h4 className="font-medium text-stone-900">{problem.title}</h4>
-                            <div className="flex gap-3 mt-1 text-xs text-stone-600">
-                              <span>Tempo: {problem.timeLimitMs}ms</span>
-                              <span>Memória: {problem.memoryLimitKb}KB</span>
-                            </div>
-                          </div>
-                        </div>
-                        <Button
-                          type="button"
-                          variant="outline"
-                          size="sm"
-                          onClick={(e) => {
-                            e.stopPropagation();
-                            onViewProblem(problem);
-                          }}
-                          className="ml-2"
-                        >
-                          Visualizar
-                        </Button>
-                      </div>
-                    </div>
-                  ))
-                )}
-              </div>
-
-              {selectedProblem && (
-                <div className="mt-3 p-3 bg-teal-50 border border-teal-200 rounded-lg">
-                  <p className="text-sm text-teal-900">
-                    <strong>Problema selecionado:</strong> {selectedProblem.title}
-                  </p>
-                </div>
-              )}
-            </div>
-
-            <div>
-              <Label htmlFor="due-date-edit">Data de Entrega *</Label>
-              <DateTimePicker
-                selected={dueDate}
-                onChange={setDueDate}
-                placeholderText="dd/mm/aaaa às hh:mm"
-                required
-                className="mt-2"
-              />
-            </div>
-
-            <div className="flex gap-3 pt-4">
-              <button
-                type="button"
-                onClick={onClose}
-                className="flex-1 px-4 py-2 border border-stone-300 rounded-lg text-stone-700 hover:bg-stone-50 font-medium transition-colors"
-              >
-                Cancelar
-              </button>
-              <button
-                type="submit"
-                disabled={!selectedProblemId || !dueDate}
-                className="flex-1 px-4 py-2 text-white rounded-lg hover:opacity-90 font-medium transition-opacity disabled:opacity-50 disabled:cursor-not-allowed" style={{ background: "linear-gradient(135deg, #0d9488 0%, #065f46 100%)" }}
-              >
-                Atualizar Atividade
-              </button>
-            </div>
-          </form>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-interface DeleteActivityModalProps {
-  isOpen: boolean;
-  onClose: () => void;
-  onConfirm: () => void;
-  activityTitle: string;
-}
-
-function DeleteActivityModal({ isOpen, onClose, onConfirm, activityTitle }: DeleteActivityModalProps) {
-  if (!isOpen) return null;
-
-  return (
-    <div className="fixed inset-0 bg-black/30 backdrop-blur-sm flex items-center justify-center z-50 p-4">
-      <div className="bg-white rounded-2xl shadow-xl max-w-md w-full p-6">
-        <div className="flex items-center gap-3 mb-4">
-          <div className="p-2 bg-red-100 rounded-lg">
-            <Trash2 className="w-6 h-6 text-red-600" />
-          </div>
-          <h3 className="text-xl font-bold text-stone-900">Confirmar Exclusão</h3>
-        </div>
-        <p className="text-stone-600 mb-6">
-          Tem certeza que deseja excluir a atividade{" "}
-          <span className="font-semibold text-stone-900">"{activityTitle}"</span>?
-          Esta ação não pode ser desfeita.
-        </p>
-        <div className="flex gap-3">
-          <button
-            onClick={onClose}
-            className="flex-1 px-4 py-2 border border-stone-300 rounded-lg text-stone-700 hover:bg-stone-50 font-medium transition-colors"
-          >
-            Cancelar
-          </button>
-          <button
-            onClick={onConfirm}
-            className="flex-1 px-4 py-2 bg-red-600 text-white rounded-lg hover:bg-red-700 font-medium transition-colors"
-          >
-            Confirmar Exclusão
-          </button>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-// Tipo para submissões dos alunos
-interface StudentSubmission {
-  studentId: number;
-  studentName: string;
-  submissionDate: string | null;
-  status: SubmissionStatus;
-  submissionId?: number; // ID da submissão específica
-  code?: string; // Código da submissão
-  language?: string; // Linguagem de programação
-}
-
-// Configuração de status para submissões
-const submissionStatusConfig = {
-  passed: {
-    label: "Aprovado",
-    icon: CheckCircle2,
-    className: "bg-green-100 text-green-800 border-green-200",
-    dotColor: "bg-green-500",
-  },
-  failed: {
-    label: "Resposta Errada",
-    icon: XCircle,
-    className: "bg-red-100 text-red-800 border-red-200",
-    dotColor: "bg-red-500",
-  },
-  pending: {
-    label: "Pendente",
-    icon: Clock,
-    className: "bg-yellow-100 text-yellow-800 border-yellow-200",
-    dotColor: "bg-yellow-500",
-  },
-  processing: {
-    label: "Processando",
-    icon: Loader2,
-    className: "bg-teal-100 text-teal-800 border-teal-200",
-    dotColor: "bg-teal-500",
-  },
-  "compile-error": {
-    label: "Erro de Compilação",
-    icon: AlertCircle,
-    className: "bg-orange-100 text-orange-800 border-orange-200",
-    dotColor: "bg-orange-500",
-  },
-  timeout: {
-    label: "Tempo Excedido",
-    icon: Clock,
-    className: "bg-teal-100 text-teal-800 border-teal-200",
-    dotColor: "bg-teal-500",
-  },
-  "runtime-error": {
-    label: "Erro de Execução",
-    icon: AlertCircle,
-    className: "bg-pink-100 text-pink-800 border-pink-200",
-    dotColor: "bg-pink-500",
-  },
-  "internal-error": {
-    label: "Erro Interno",
-    icon: XCircle,
-    className: "bg-stone-100 text-stone-800 border-stone-200",
-    dotColor: "bg-stone-500",
-  },
-  unknown: {
-    label: "Desconhecido",
-    icon: AlertCircle,
-    className: "bg-stone-100 text-stone-800 border-stone-200",
-    dotColor: "bg-stone-500",
-  },
-} as const;
-
-interface SubmissionStatusBadgeProps {
-  status: SubmissionStatus;
-}
-
-function SubmissionStatusBadge({ status }: SubmissionStatusBadgeProps) {
-  const config = submissionStatusConfig[status] || submissionStatusConfig.unknown;
-  const Icon = config.icon;
-
-  return (
-    <span
-      className={`inline-flex items-center gap-1.5 px-2.5 py-1 rounded-full text-xs font-medium border ${config.className}`}
-    >
-      <div className={`w-1.5 h-1.5 rounded-full ${config.dotColor}`} />
-      <Icon className="w-3 h-3" />
-      {config.label}
-    </span>
-  );
-}
-
-interface SubmissionsModalProps {
-  isOpen: boolean;
-  onClose: () => void;
-  activity: Activity | null;
-  classId: number;
-}
-
-function SubmissionsModal({ isOpen, onClose, activity, classId }: SubmissionsModalProps) {
-  const [submissions, setSubmissions] = useState<StudentSubmission[]>([]);
-  const [loading, setLoading] = useState(false);
-  const [selectedSubmission, setSelectedSubmission] = useState<{
-    code: string;
-    submission: Submission;
-    studentName: string;
-  } | null>(null);
-  const [loadingDetails, setLoadingDetails] = useState(false);
-  const [testResults, setTestResults] = useState<any[]>([]);
-  const [problem, setProblem] = useState<Problem | null>(null);
-
-  useEffect(() => {
-    if (isOpen && activity) {
-      loadSubmissions();
-      loadProblem();
-    }
-  }, [isOpen, activity]);
-
-  const loadProblem = async () => {
-    if (!activity) return;
-
-    try {
-      const problemData = await getProblemById(String(activity.problemId));
-      setProblem(problemData);
-    } catch (error) {
-      console.error("Erro ao carregar problema:", error);
-    }
-  };
-
-  const loadSubmissions = async () => {
-    if (!activity) return;
-
-    setLoading(true);
-    try {
-      const data = await getActivitySubmissions(classId, activity.id);
-
-      // Agrupar submissões por usuário e aplicar lógica de priorização
-      const submissionsByUser = new Map<number, any>();
-
-      data.forEach((item: any) => {
-        const userId = item.user_id;
-        const existing = submissionsByUser.get(userId);
-
-        if (!existing) {
-          // Se não existe submissão deste usuário, adiciona
-          submissionsByUser.set(userId, item);
-        } else {
-          const isCurrentAccepted = item.status === "Aceita";
-          const isExistingAccepted = existing.status === "Aceita";
-          const currentDate = new Date(item.created_at);
-          const existingDate = new Date(existing.created_at);
-
-          // Lógica de priorização:
-          // 1. Se a atual é Aceita e a existente não é → substitui
-          // 2. Se ambas são Aceita → pega a mais recente
-          // 3. Se nenhuma é Aceita → pega a mais recente
-          // 4. Se a existente é Aceita e a atual não é → mantém a existente
-
-          if (isCurrentAccepted && !isExistingAccepted) {
-            submissionsByUser.set(userId, item);
-          } else if (isCurrentAccepted && isExistingAccepted && currentDate > existingDate) {
-            submissionsByUser.set(userId, item);
-          } else if (!isCurrentAccepted && !isExistingAccepted && currentDate > existingDate) {
-            submissionsByUser.set(userId, item);
-          }
-          // Caso contrário, mantém a existente (não faz nada)
-        }
-      });
-
-      // Mapear os dados da API para o formato esperado pelo componente
-      const mappedSubmissions: StudentSubmission[] = Array.from(submissionsByUser.values()).map((item: any) => ({
-        studentId: item.user_id,
-        studentName: item.user_name,
-        submissionDate: item.created_at,
-        status: mapStatusToSubmissionStatus(item.status),
-        submissionId: item.id, // Adicionar ID da submissão
-        code: item.codigo, // Adicionar código da submissão
-        language: item.linguagem === 50 ? 'c' : 'c', // Mapear linguagem (50 = C)
-      }));
-
-      mappedSubmissions.sort((a, b) => a.studentName.localeCompare(b.studentName));
-
-      setSubmissions(mappedSubmissions);
-    } catch (error) {
-      console.error("Erro ao carregar submissões:", error);
-      setSubmissions([]);
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  // Função auxiliar para mapear status da API para SubmissionStatus
-  const mapStatusToSubmissionStatus = (status: string): SubmissionStatus => {
-    const statusMap: Record<string, SubmissionStatus> = {
-      "Aceita": "passed",
-      "Resposta Errada": "failed",
-      "Erro de Compilação": "compile-error",
-      "Erro de Execução (NZEC)": "runtime-error",
-      "Tempo Limite Excedido": "timeout",
-      "Erro Interno": "internal-error",
-      "Processando": "processing",
-      "Pendente": "pending",
-    };
-
-    return statusMap[status] || "unknown";
-  };
-
-  const handleRowClick = async (submission: StudentSubmission & { submissionId?: number }) => {
-    if (!submission.submissionId || !submission.code) return;
-
-    setLoadingDetails(true);
-    try {
-      // Usar o código que já está armazenado na submissão
-      setSelectedSubmission({
-        code: submission.code,
-        submission: {
-          id: submission.submissionId,
-          activityId: activity!.id,
-          dateSubmitted: submission.submissionDate || new Date().toISOString(),
-          language: (submission.language || 'c') as any,
-          status: submission.status,
-          problemTitle: null,
-        },
-        studentName: submission.studentName,
-      });
-
-      // Carregar resultados dos testes
-      const results = await getResultBySubmissionId(submission.submissionId);
-      setTestResults(results);
-    } catch (error) {
-      console.error("Erro ao carregar detalhes da submissão:", error);
-    } finally {
-      setLoadingDetails(false);
-    }
-  };
-
-  const handleCloseDetails = () => {
-    setSelectedSubmission(null);
-    setTestResults([]);
-  };
-
-  if (!isOpen || !activity) return null;
-
-  // Visualização dos detalhes da submissão
-  if (selectedSubmission) {
-    return (
-      <div className="fixed inset-0 bg-black/30 backdrop-blur-sm flex items-center justify-center z-50 p-4">
-        <div className="bg-white rounded-2xl shadow-xl max-w-6xl w-full max-h-[90vh] overflow-hidden flex flex-col">
-          <div className="p-6 border-b border-stone-200">
-            <div className="flex items-center justify-between">
-              <div>
-                <h2 className="text-xl font-bold text-stone-900">
-                  Submissão de {selectedSubmission.studentName}
-                </h2>
-                <p className="text-sm text-stone-500 mt-1">
-                  {new Date(selectedSubmission.submission.dateSubmitted).toLocaleString("pt-BR")}
-                </p>
-              </div>
-              <button
-                onClick={handleCloseDetails}
-                className="text-stone-400 hover:text-stone-600 transition-colors"
-              >
-                <X className="w-5 h-5" />
-              </button>
-            </div>
-          </div>
-
-          <div className="flex-1 overflow-y-auto p-6 space-y-6">
-            {/* Código submetido */}
-            <CodeViewer code={selectedSubmission.code} language={selectedSubmission.submission.language} />
-
-            {/* Casos de teste */}
-            {problem && problem.testCases && problem.testCases.length > 0 && (
-              <div className="bg-white rounded-lg border border-stone-200 shadow-sm">
-                <div className="border-b border-stone-200 p-6">
-                  <div className="flex items-center gap-3">
-                    <div className="p-2 bg-green-100 rounded-lg">
-                      <TestTube className="w-5 h-5 text-green-600" />
-                    </div>
-                    <h2 className="text-2xl font-bold text-stone-900">
-                      Casos de Teste
-                    </h2>
-                  </div>
-                </div>
-
-                <div className="p-6">
-                  <div className="overflow-hidden">
-                    <Table>
-                      <TableHeader>
-                        <TableRow className="bg-stone-50 hover:bg-stone-50">
-                          <TableHead className="font-semibold text-stone-900">
-                            <div className="flex items-center gap-2">
-                              <Hash className="w-4 h-4" />
-                              Teste
-                            </div>
-                          </TableHead>
-                          <TableHead className="font-semibold text-stone-900">
-                            <div className="flex items-center gap-2">
-                              <Clock className="w-4 h-4" />
-                              Status
-                            </div>
-                          </TableHead>
-                          <TableHead className="font-semibold text-stone-900">
-                            <div className="flex items-center gap-2">
-                              <Terminal className="w-4 h-4" />
-                              Saída Atual
-                            </div>
-                          </TableHead>
-                          <TableHead className="font-semibold text-stone-900">
-                            <div className="flex items-center gap-2">
-                              <TargetIcon className="w-4 h-4" />
-                              Saída Esperada
-                            </div>
-                          </TableHead>
-                        </TableRow>
-                      </TableHeader>
-                      <TableBody>
-                        {problem.testCases.map((testCase, index) => {
-                          const result = testResults.find(r => r.testCaseId === testCase.id);
-                          return (
-                            <TestCaseRow
-                              key={testCase.id}
-                              testCase={testCase}
-                              result={result}
-                              index={index}
-                            />
-                          );
-                        })}
-                      </TableBody>
-                    </Table>
-                  </div>
-                </div>
-              </div>
-            )}
-          </div>
-
-          <div className="p-6 border-t border-stone-200 flex justify-end">
-            <Button onClick={handleCloseDetails} variant="outline">
-              Voltar para lista
-            </Button>
-          </div>
-        </div>
-      </div>
-    );
-  }
-
-  // Lista de submissões
-  return (
-    <div className="fixed inset-0 bg-black/30 backdrop-blur-sm flex items-center justify-center z-50 p-4">
-      <div className="bg-white rounded-2xl shadow-xl max-w-5xl w-full max-h-[90vh] overflow-hidden flex flex-col">
-        <div className="p-6 border-b border-stone-200">
-          <div className="flex items-center justify-between">
-            <h2 className="text-xl font-bold text-stone-900">Submissões da Atividade</h2>
-            <button
-              onClick={onClose}
-              className="text-stone-400 hover:text-stone-600 transition-colors"
-            >
-              <X className="w-5 h-5" />
-            </button>
-          </div>
-        </div>
-
-        <div className="flex-1 overflow-y-auto p-6">
-          {loading ? (
-            <div className="flex items-center justify-center py-12">
-              <Loading />
-            </div>
-          ) : submissions.length === 0 ? (
-            <div className="text-center py-12">
-              <Users className="w-12 h-12 text-stone-400 mx-auto mb-4" />
-              <h3 className="text-lg font-medium text-stone-900 mb-2">
-                Nenhuma submissão encontrada
-              </h3>
-              <p className="text-stone-500">
-                Ainda não há submissões para esta atividade
-              </p>
-            </div>
-          ) : (
-            <div className="bg-white rounded-lg border border-stone-200 shadow-sm overflow-hidden">
-              <Table>
-                <TableHeader>
-                  <TableRow className="bg-stone-50 hover:bg-stone-50">
-                    <TableHead className="font-semibold text-stone-900">
-                      <div className="flex items-center gap-2">
-                        <Users className="w-4 h-4" />
-                        Nome
-                      </div>
-                    </TableHead>
-                    <TableHead className="font-semibold text-stone-900">
-                      <div className="flex items-center gap-2">
-                        <Calendar className="w-4 h-4" />
-                        Data de Submissão
-                      </div>
-                    </TableHead>
-                    <TableHead className="font-semibold text-stone-900">
-                      <div className="flex items-center gap-2">
-                        <Clock className="w-4 h-4" />
-                        Status
-                      </div>
-                    </TableHead>
-                  </TableRow>
-                </TableHeader>
-                <TableBody>
-                  {submissions.map((submission) => (
-                    <TableRow
-                      key={submission.studentId}
-                      onClick={() => handleRowClick(submission as StudentSubmission & { submissionId?: number })}
-                      className="cursor-pointer hover:bg-teal-50 transition-colors duration-200 group"
-                    >
-                      <TableCell className="font-medium">
-                        <span className="text-stone-900 group-hover:text-teal-600 transition-colors">
-                          {submission.studentName}
-                        </span>
-                      </TableCell>
-                      <TableCell>
-                        {submission.submissionDate ? (
-                          <div className="flex flex-col">
-                            <span className="text-stone-900 font-medium">
-                              {new Date(submission.submissionDate).toLocaleDateString("pt-BR")}
-                            </span>
-                            <span className="text-xs text-stone-500">
-                              {new Date(submission.submissionDate).toLocaleTimeString("pt-BR", {
-                                hour: "2-digit",
-                                minute: "2-digit",
-                                second: "2-digit",
-                              })}
-                            </span>
-                          </div>
-                        ) : (
-                          <span className="text-stone-400 text-sm">Não submetido</span>
-                        )}
-                      </TableCell>
-                      <TableCell>
-                        <SubmissionStatusBadge status={submission.status} />
-                      </TableCell>
-                    </TableRow>
-                  ))}
-                </TableBody>
-              </Table>
-            </div>
-          )}
-
-          {loadingDetails && (
-            <div className="fixed inset-0 bg-black/20 flex items-center justify-center z-50">
-              <div className="bg-white rounded-lg p-6 shadow-xl">
-                <Loader2 className="w-8 h-8 animate-spin text-teal-600 mx-auto mb-2" />
-                <p className="text-stone-600">Carregando detalhes...</p>
-              </div>
-            </div>
-          )}
-        </div>
-      </div>
-    </div>
-  );
-}
-
-// Componente auxiliar para linha de caso de teste
-interface TestCaseRowProps {
-  testCase: any;
-  result: any;
-  index: number;
-}
-
-function TestCaseRow({ testCase, result, index }: TestCaseRowProps) {
-  const [showOutput, setShowOutput] = useState(false);
-
-  const actualOutput = result?.stdout || result?.stderr || "Sem saída";
-  const expectedOutput = testCase.expectedOutput || "Sem saída esperada";
-
-  const copyToClipboard = (text: string) => {
-    navigator.clipboard.writeText(text);
-  };
-
-  return (
-    <TableRow className="hover:bg-stone-50 transition-colors duration-200">
-      <TableCell className="font-medium">
-        <div className="flex items-center gap-2">
-          <Hash className="w-4 h-4 text-stone-400" />
-          <span>Teste {index + 1}</span>
-        </div>
-      </TableCell>
-      <TableCell>
-        <SubmissionStatusBadge status={(result?.status || 'pending') as SubmissionStatus} />
-      </TableCell>
-      <TableCell>
-        <div className="space-y-2">
-          <div className="flex items-center gap-2">
-            <Button
-              variant="ghost"
-              size="sm"
-              onClick={() => setShowOutput(!showOutput)}
-              className="h-6 px-2 text-xs"
-            >
-              {showOutput ? (
-                <EyeOff className="w-3 h-3" />
-              ) : (
-                <Eye className="w-3 h-3" />
-              )}
-              {showOutput ? "Ocultar" : "Mostrar"}
-            </Button>
-            {actualOutput && (
-              <Button
-                variant="ghost"
-                size="sm"
-                onClick={() => copyToClipboard(actualOutput)}
-                className="h-6 px-2 text-xs"
-              >
-                <Copy className="w-3 h-3" />
-              </Button>
-            )}
-          </div>
-          {showOutput && (
-            <div className="bg-stone-900 rounded p-2 max-w-xs">
-              <pre className="text-xs text-stone-300 whitespace-pre-wrap overflow-auto max-h-20">
-                {actualOutput}
-              </pre>
-            </div>
-          )}
-        </div>
-      </TableCell>
-      <TableCell>
-        <div className="space-y-2">
-          <div className="flex items-center gap-2">
-            <Button
-              variant="ghost"
-              size="sm"
-              onClick={() => copyToClipboard(expectedOutput)}
-              className="h-6 px-2 text-xs"
-            >
-              <Copy className="w-3 h-3" />
-              Copiar
-            </Button>
-          </div>
-          <div className="bg-stone-900 rounded p-2 max-w-xs">
-            <pre className="text-xs text-stone-300 whitespace-pre-wrap overflow-auto max-h-20">
-              {expectedOutput}
-            </pre>
-          </div>
-        </div>
-      </TableCell>
-    </TableRow>
   );
 }
